Refactored script for processing delays
